[remove]remove Selection class
using Selected component instead.
