+++ game_editor/targets/windows/main/CMakeLists.txt
@@ -5,13 +5,15 @@ add_library(${PROJECT_NAME} STATIC)
 target_sources(${PROJECT_NAME} PRIVATE
     src/Editor.cpp
     src/WinMain.cpp
-    src/scene_gizmo_renderer.cpp
+    src/component_editors/physics_shape_editor.cpp
+    src/component_editors/prefab_editor.cpp
+    src/editor_windows/entity_inspector_window.cpp
+    src/editor_windows/material_editor_window.cpp
     src/editor_windows/scene_hierarchy_window.cpp
     src/editor_windows/scene_view_window.cpp
-    src/editor_windows/material_editor_window.cpp
-    src/component_editors/prefab_editor.cpp
-    src/component_editors/physics_shape_editor.cpp
+    src/scene_gizmo_renderer.cpp
 )
+
 target_link_libraries(${PROJECT_NAME}
     PRIVATE
     nodec_game_engine_core
