+++ game_editor/targets/windows/main/src/editor_windows/scene_hierarchy_window.hpp
@@ -19,10 +19,6 @@ public:
 
     void on_gui() override;
 
-    decltype(auto) selected_entity_changed() {
-        return selected_entity_changed_.signal_interface();
-    }
-
 private:
     void show_entity_node(const nodec_scene::SceneEntity entity);
 
@@ -33,9 +29,6 @@ private:
     nodec_scene::Scene &scene_;
 
     nodec_scene::SceneEntity selected_entity_{nodec::entities::null_entity};
-
-    nodec::signals::Signal<void(nodec_scene::SceneEntity selected)> selected_entity_changed_;
-
     std::unique_ptr<nodec_scene_serialization::SerializableEntity> copied_entity_;
 };
